refactor: Sections 일급 컬렉션 사용
