+++ src/main/java/nextstep/subway/domain/Line.java
@@ -10,15 +10,13 @@ import java.util.List;
 
 @Entity
 public class Line extends BaseEntity {
-    private static final int MIN_SECTIONS_SIZE = 1;
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
     private String color;
-    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
-    private List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections = new Sections();
 
     protected Line() {
     }
@@ -45,65 +43,12 @@ public class Line extends BaseEntity {
         this.color = color;
     }
 
-    public void removeSection(Long stationId) {
-        validateSectionCount();
-        validateIsLastStation(stationId);
-        sections.remove(getLastSection());
-    }
-
-    private void validateIsLastStation(Long stationId) {
-        if (!isLastStation(stationId)) {
-            throw new RemoveSectionFailException();
-        }
-    }
-
-    private void validateSectionCount() {
-        if (sections.size() == MIN_SECTIONS_SIZE) {
-            throw new RemoveSectionFailException();
-        }
-    }
-
-    private boolean isLastStation(Long stationId) {
-        return getLastDownStation().getId()
-                .equals(stationId);
-    }
-
     public void addSection(Section section) {
-        if (!sections.isEmpty()) {
-            validateNewSection(section);
-        }
-
-        this.sections.add(section);
         section.setLine(this);
+        sections.addSection(section);
     }
 
-    private void validateNewSection(Section section) {
-        validateUpStation(section.getUpStation());
-        validateDownStation(section.getDownStation());
-    }
-
-    private void validateUpStation(Station upStation) {
-        boolean isConnectedStation = getLastDownStation().equals(upStation);
-
-        if (!isConnectedStation) {
-            throw new InvalidUpStationException(upStation.getName());
-        }
-    }
-
-    private Station getLastDownStation() {
-        return getLastSection().getDownStation();
-    }
-
-    private Section getLastSection() {
-        return sections.get(sections.size() - 1);
-    }
-
-    private void validateDownStation(Station downStation) {
-        sections.stream()
-                .filter(s -> s.contains(downStation))
-                .findAny()
-                .ifPresent(s -> {
-                    throw new InvalidDownStationException(downStation.getName());
-                });
+    public void removeSection(Long stationId) {
+        sections.removeSection(stationId);
     }
 }
